Downgrade fuel-core-client to 0.17.2; indexers need to be fixed

+++ packages/fuel-indexer-database/src/queries.rs
@@ -264,7 +264,7 @@ pub async fn last_block_height_for_indexer(
     conn: &mut IndexerConnection,
     namespace: &str,
     identifier: &str,
-) -> sqlx::Result<u32> {
+) -> sqlx::Result<u64> {
     match conn {
         IndexerConnection::Postgres(ref mut c) => {
             postgres::last_block_height_for_indexer(c, namespace, identifier).await
